Migrate to new Warrior image, other enhancements
This change detects the old Warrior image, backs up the user config.json, deletes the container, and then prunes everything from Docker. If the new image or the Watchtower image aren't found, they are installed and configured. The user configuration is now stored on the host system and bind mounted into the Docker image so it will survive container updates as well as the complete deletion of the container. Additionally, container names are used instead of container IDs when referring to containers since Watchtower doesn't preserve container IDs on automatic updates.